chore(package): change library name from SPARQL2GraphQL to SPARQLess
