+++ README.md
@@ -1,14 +1,14 @@
-# SPARQL2GrahpQL
+# SPARQLess
 
 Many government and non-government organizations today publish various datasets using Linked Data through SPARQL endpoints.
 These datasets are often also available as JSON or RDF dumps, but having to write code to explore them can be cumbersome.
-SPARQL2GraphQL aims to make it easier to use these datasets by providing a tool which will give developers a more
-friendly interface which they likely already know how to use - GraphQL.
+SPARQLess (/_spark less_/) aims to make it easier to use these datasets by providing a tool which will give developers a more
+friendly interface which they likely already know how to use - [GraphQL](https://graphql.org/learn/).
 
-Simply configure SPARQL2GraphQL with the URL of a SPARQL endpoint which has some data which interests you, and
+Simply configure SPARQLess with the URL of a SPARQL endpoint which has some data which interests you, and
 let it do its magic. Soon, you will have a GraphQL instance which you can use to painlessly explore and query the data.
 
 ## Documentation
 
-The documentation located [here](https://yawnston.github.io/sparql2graphql/)
+The documentation located [here](https://mff-uk.github.io/sparqless/)
 includes installation instructions, programmer and user guides, and much more.
