+++ src/api/builder.ts
@@ -1,0 +1,3 @@
+export class SparqlessBuilder {
+    // TODO: implement a builder API
+}
